feat: creates a wrapper for the TabView.
It is not possible to use a template for the TabViewItems when using WebView (the webview2 control is not created "in time").
And making a control inherit from TabViewItem is discouraged, in practice it caused issues.

This solution is easier to maintain and is similar to what we have with the WebView2 wrapper.